requirements.txt
django4.1.3
requests==2.28.1
Pillow==9.2.0
mysqlclient==2.2.0
python-decouple==3.4
python-dotenv==0.21.0